- changed metric default aggregate behavior to store, instead of average, to ensure backwards compatibility over all inherited metrics
